claude: refactor(kaspa): move relayer module from libs/kaspa to dymension-kaspa
Inverts dependency hierarchy by moving relayer code into dymension-kaspa chain implementation.

Changes:
- Moved dymension/libs/kaspa/lib/relayer/src/ to rust/main/chains/dymension-kaspa/src/kas_relayer/
- Renamed lib.rs to mod.rs to reflect module structure
- Updated all imports from corelib/hardcode to dym_kas_core/dym_kas_hardcode
- Fixed relative imports within kas_relayer module
- Removed relayer from libs/kaspa workspace
- Removed relayer dependency from validator and tooling crates
- Updated dymension-kaspa Cargo.toml to add secp256k1 and dym-kas-hardcode dependencies
- Updated hyperlane-base to use dymension_kaspa::kas_relayer instead of dym_kas_relayer

Result: libs/kaspa now has no hyperlane dependencies, enabling cleaner separation.

Closes #140
