--- dymension/libs/kaspa/lib/relayer/Cargo.toml
@@ -1,43 +1,0 @@
-[package]
-name = "relayer"
-version = "0.2.0"
-edition = "2021"
-
-[dependencies]
-corelib = { path = "../core", package = "core" }
-hardcode = { path = "../hardcode", package = "hardcode" }
-kaspa-addresses = { workspace = true }
-kaspa-consensus-core = { workspace = true }
-kaspa-core = { workspace = true }
-kaspa-grpc-client = { workspace = true }
-kaspa-hashes = { workspace = true }
-kaspa-wallet-core = { workspace = true }
-kaspa-wallet-pskt = { workspace = true }
-kaspa-txscript = { workspace = true }
-kaspa-rpc-core = { workspace = true }
-kaspa-consensus-client = { workspace = true }
-
-rand = "0.8"   # Use the latest stable version
-secp256k1 = { workspace = true }
-workflow-core = { workspace = true }
-tokio = { workspace = true }
-log = { workspace = true }
-tracing = { workspace = true }
-reqwest = { workspace = true }
-eyre = { workspace = true }
-bytes = { workspace = true }
-anyhow = "1.0"
-base64 = "0.22.0"
-hex = { workspace = true }
-api-rs = { path = "../api", package = "openapi" }
-hyperlane-core = { path = "../../../../../rust/main/hyperlane-core" }
-hyperlane-warp-route = { path = "../../../../../rust/main/applications/hyperlane-warp-route" }
-hyperlane-cosmos = { path = "../../../../../rust/main/chains/hyperlane-cosmos" }
-serde = { version = "^1.0", features = ["derive"] }
-prost = { version = "0.13", default-features = false }
-hyperlane-cosmos-rs = { workspace = true }
-reqwest-middleware = "0.4"
-
-
-bincode = "1.3"
-prometheus = "0.13"
