--- dymension/libs/kaspa/lib/relayer/README.md
@@ -1,3 +1,0 @@
-## What?
-
-Contains logic and types used by relayer but not validator.
