+++ dymension/libs/kaspa/tooling/Cargo.toml
@@ -8,7 +8,6 @@ kaspa-core = { workspace = true }
 corelib = { path = "../lib/core", package = "core" }
 hardcode = { path = "../lib/hardcode" }
 validator = { path = "../lib/validator" }
-relayer = { path = "../lib/relayer" }
 secp256k1 = { workspace = true }
 serde = { version = "^1.0", features = ["derive"] }
 serde_json = { workspace = true }
