+++ rust/main/chains/dymension-kaspa/Cargo.toml
@@ -50,14 +50,15 @@ url = { workspace = true }
 reqwest = { workspace = true }
 axum = { workspace = true }
 dym-kas-validator = { path = "../../../../dymension/libs/kaspa/lib/validator", package="validator"}
-dym-kas-relayer = { path = "../../../../dymension/libs/kaspa/lib/relayer" , package="relayer"}
 dym-kas-core = { path = "../../../../dymension/libs/kaspa/lib/core" , package="core"}
+dym-kas-hardcode = { path = "../../../../dymension/libs/kaspa/lib/hardcode", package="hardcode"}
 dym-kas-api = { path = "../../../../dymension/libs/kaspa/lib/api" , package="openapi"}
 dym-kas-kms = { path = "../../../../dymension/libs/kaspa/lib/kms", package="dym-kas-kms"}
 dymension-kaspa-hl-constants = { path = "../dymension-kaspa-hl-constants" }
 eyre = { workspace = true }
 bytes = { workspace = true }
 prometheus = { workspace = true }
+secp256k1 = { version = "0.29.0", features = ["global-context", "rand-std", "serde"] }
 
 hyperlane-metric = { path = "../../hyperlane-metric" }
 hyperlane-core = { path = "../../hyperlane-core", features = ["async"] }
