+++ rust/main/chains/dymension-kaspa/src/kas_relayer/deposit/deposit.rs
@@ -1,9 +1,8 @@
-use corelib::api::client::{Deposit, HttpClient};
-use corelib::deposit::DepositFXG;
-use corelib::finality::is_safe_against_reorg;
+use dym_kas_core::api::client::{Deposit, HttpClient};
+use dym_kas_core::deposit::DepositFXG;
+use dym_kas_core::finality::is_safe_against_reorg;
 use eyre::Result;
 use hyperlane_core::U256;
-pub use secp256k1::PublicKey;
 use tracing::{debug, info};
 
 /// Error type for deposit processing that includes retry timing information
@@ -119,7 +118,7 @@ pub fn build_deposit_fxg(
 #[cfg(test)]
 mod tests {
 
-    use corelib::message::ParsedHL;
+    use dym_kas_core::message::ParsedHL;
 
     #[test]
     fn test_parsed_hl_parse() {
