+++ rust/main/chains/dymension-kaspa/src/kas_relayer/mod.rs
@@ -8,5 +8,3 @@ pub use withdraw::messages::on_new_withdrawals;
 
 // Re-export metrics for easier access
 pub use metrics::KaspaBridgeMetrics;
-
-pub use secp256k1::PublicKey;
