+++ rust/main/chains/dymension-kaspa/src/kas_relayer/withdraw/sweep.rs
@@ -1,11 +1,11 @@
-use crate::withdraw::messages::PopulatedInput;
-use crate::withdraw::populated_input::PopulatedInputBuilder;
-use corelib::consts::RELAYER_SIG_OP_COUNT;
-use corelib::escrow::EscrowPublic;
-use corelib::util::input_sighash_type;
-use corelib::wallet::EasyKaspaWallet;
+use super::messages::PopulatedInput;
+use super::populated_input::PopulatedInputBuilder;
+use dym_kas_core::consts::RELAYER_SIG_OP_COUNT;
+use dym_kas_core::escrow::EscrowPublic;
+use dym_kas_core::util::input_sighash_type;
+use dym_kas_core::wallet::EasyKaspaWallet;
 use eyre::{eyre, Result};
-use hardcode::tx::{DUST_AMOUNT, RELAYER_SWEEPING_PRIORITY_FEE};
+use dym_kas_hardcode::tx::{DUST_AMOUNT, RELAYER_SWEEPING_PRIORITY_FEE};
 use kaspa_consensus_client::{
     TransactionOutpoint as ClientTransactionOutpoint, UtxoEntry as ClientUtxoEntry,
 };
