+++ rust/main/chains/dymension-kaspa/src/providers/provider.rs
@@ -9,9 +9,9 @@ use crate::ValidatorStuff;
 use dym_kas_core::confirmation::ConfirmationFXG;
 use dym_kas_core::escrow::EscrowPublic;
 use dym_kas_core::wallet::{EasyKaspaWallet, EasyKaspaWalletArgs};
-use dym_kas_relayer::withdraw::hub_to_kaspa::combine_bundles_with_fee;
-use dym_kas_relayer::withdraw::messages::on_new_withdrawals;
-use dym_kas_relayer::KaspaBridgeMetrics;
+use crate::kas_relayer::withdraw::hub_to_kaspa::combine_bundles_with_fee;
+use crate::kas_relayer::withdraw::messages::on_new_withdrawals;
+use crate::kas_relayer::KaspaBridgeMetrics;
 use eyre::Result;
 use hyperlane_core::config::OpSubmissionConfig;
 use hyperlane_core::NativeToken;
