+++ rust/main/hyperlane-base/Cargo.toml
@@ -73,7 +73,6 @@ hyperlane-radix = { path = "../chains/hyperlane-radix" }
 dym_kas_core = { path = "../../../dymension/libs/kaspa/lib/core" , package="core"}
 dym_kas_hardcode = { path = "../../../dymension/libs/kaspa/lib/hardcode" , package="hardcode"}
 api-rs = { path = "../../../dymension/libs/kaspa/lib/api", package = "openapi" }
-dym_kas_relayer = { path = "../../../dymension/libs/kaspa/lib/relayer" , package="relayer"}
 dym_kas_validator = { path = "../../../dymension/libs/kaspa/lib/validator" , package="validator"}
 
 kaspa-consensus-core = { git = "https://github.com/kaspanet/rusty-kaspa", rev = "9ff5d0f" }
