+++ rust/main/hyperlane-base/src/kas_hack/logic_loop.rs
@@ -3,8 +3,8 @@ use dymension_kaspa::hl_message::add_kaspa_metadata_hl_messsage;
 use dym_kas_core::{
     confirmation::ConfirmationFXG, deposit::DepositFXG, finality::is_safe_against_reorg,
 };
-use dym_kas_relayer::confirm::expensive_trace_transactions;
-use dym_kas_relayer::deposit::{build_deposit_fxg, check_deposit_finality, KaspaTxError};
+use dymension_kaspa::kas_relayer::confirm::expensive_trace_transactions;
+use dymension_kaspa::kas_relayer::deposit::{build_deposit_fxg, check_deposit_finality, KaspaTxError};
 use dymension_kaspa::{Deposit, KaspaProvider};
 use ethers::utils::hex::ToHex;
 use eyre::Result;
